Save and Load
Created the save file function, and made the loadfile function check if a file md_{filename}.m exists or not. If exists - prompt user to select to use it or to use a md_defaults while loading file, or both (use md from file md and rest from md_defaults).