Updated GUI
Cleaned up the GUI callback def function - created new functions with the functions previously in the callback def.

Reconstructed the edit plot conf function; put the old one in _old.

Made the open variables function check if it is a user defined or built in plot conf of which the variables are to be opened.

Minor updates and bug fixes otherwise.